change from extending subclass to using mixin, fix inline form issue if a form field is the last 'word'
